+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ## [Unreleased]
 
 - Update `irx-config`
+- Fix register array derive regression
 
 ## [v0.36.0] - 2025-03-09